Widen function indices from 8 to 16 bits, raise DIAMOND_MAX_FUNCTIONS to 512
The 256-function cap was a real architectural ceiling, not a struct-sizing
choice: function indices were uint8_t throughout the bytecode format
(CALL/CALL_TYPED/CLOSURE operands, DiamondMethod/DiamondClosure.
function_index), so raising the constant alone would have silently
wrapped indices past 256. The self-hosted parser's own bootstrap
self-compile was already at 251/256 after the prior Enumerable round, and
the next self-hosting phase (exceptions, modules, require) needed more
headroom than that.

Widened to big-endian 16-bit, matching the existing JUMP operand
convention exactly. The self-hosted parser needed the identical format
change on its own emission side (it hand-assembles its own bytecode via
ProgramBuilder), plus a fix to a test that hand-assembles a raw CALL
instruction byte-by-byte against the old one-byte layout.

+++ src/object.h
@@ -143,7 +143,7 @@ typedef struct DiamondHash {
 
 typedef struct DiamondClosure {
     DiamondObject object;
-    uint8_t function_index;
+    uint16_t function_index;
     uint8_t capture_count;
     DiamondValue captures[16];
 } DiamondClosure;
